move query components to Developers namespace
