--- app/components/developer_query_mobile_component.rb
@@ -1,2 +1,0 @@
-class DeveloperQueryMobileComponent < DeveloperQueryComponent
-end
