+++ app/components/developers/query_component.html.erb
