+++ app/components/developers/query_mobile_component.rb
@@ -1,0 +1,4 @@
+module Developers
+  class QueryMobileComponent < Developers::QueryComponent
+  end
+end
