+++ config/locales/en.yml
@@ -173,21 +173,6 @@ en:
   developer_profile_action_component:
     add_profile: Add my profile
     edit_profile: Update my profile
-  developer_query_component:
-    apply: Apply
-    clear: Clear
-    close_menu: Close menu
-    filters: Filters
-    include_not_interested: Include developers not currently interested
-    role_level:
-      title: Role level
-    search_query_label: Search profiles
-    search_query_placeholder: Hotwire, React, Heroku, etc.
-    timezone:
-      title: Timezone
-    title: Developer filters
-    work_preference:
-      title: Work preference
   developer_query_sort_button_component:
     sort:
       availability: Availability
@@ -259,6 +244,21 @@ en:
       description: "%{developers_count}+ Ruby on Rails developers looking for their next gig. Juniors to seniors and everyone in between, you'll find them all here."
       title: Ruby on Rails developers
       title_og: Hire Ruby on Rails developers
+    query_component:
+      apply: Apply
+      clear: Clear
+      close_menu: Close menu
+      filters: Filters
+      include_not_interested: Include developers not currently interested
+      role_level:
+        title: Role level
+      search_query_label: Search profiles
+      search_query_placeholder: Hotwire, React, Heroku, etc.
+      timezone:
+        title: Timezone
+      title: Developer filters
+      work_preference:
+        title: Work preference
     show:
       description: Description
       edit: Edit
